// src/frontend/src/api/model/genericSort.ts
/**
 * ExampleWebApp API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SortModelItem } from './sortModelItem';


export interface GenericSort { 
    columns?: Array<SortModelItem> | null;
}