// src/frontend/src/api/model/sortModelItem.ts
/**
 * ExampleWebApp API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SortDirection } from './sortDirection';


export interface SortModelItem { 
    columnName?: string | null;
    direction?: SortDirection;
}
export namespace SortModelItem {
}